design(#30): 채팅창 컴포넌트 분리
