+++ src/components/chat/ChatRoom.tsx
@@ -1,0 +1,8 @@
+import styled from 'styled-components';
+import { theme } from "@/styles/theme";
+
+const ChatRoom = () => {
+
+};
+
+export default ChatRoom;
