+++ src/components/common/ChatButton.tsx
@@ -3,6 +3,7 @@ import Image from "next/image";
 import { useEffect, useState } from "react";
 import styled from "styled-components";
 import ChatWindow from "../chat/ChatWIndow";
+import List from "../chat/List";
 
 interface msgCountProps {
   count: number;
@@ -34,7 +35,7 @@ const ChatButton = (props: msgCountProps) => {
   return (
     <>
       {isChatOpen &&
-        <ChatWindow
+        <List
           onClose={handleChatWindowClose} />}
       <MsgButton onClick={toggleChat}>
         <Image
